+++ examples/angular-router/src/app/components/search/search.component.ts
@@ -1,16 +1,10 @@
 import { Component, OnDestroy } from '@angular/core';
+import algoliasearch from 'algoliasearch/lite';
 
 @Component({
   selector: 'app-search',
   template: `
-    <ais-instantsearch
-      [config]="{
-        appId: 'latency',
-        apiKey: '6be0576ff61c053d5f9a3225e2a90f76',
-        indexName: 'instant_search',
-        routing: true
-      }"
-    >
+    <ais-instantsearch [config]="config">
       <div class="jumbotron">
         <p class="text-center">
           <ais-search-box placeholder="Search a product"></ais-search-box>
@@ -43,9 +37,9 @@ import { Component, OnDestroy } from '@angular/core';
               <ais-sort-by
                 [items]="
                   [
-                    {name: 'instant_search', label: 'Featured'},
-                    {name: 'instant_search_price_asc', label: 'Price asc.'},
-                    {name: 'instant_search_price_desc', label: 'Price desc.'}
+                    {value: 'instant_search', label: 'Featured'},
+                    {value: 'instant_search_price_asc', label: 'Price asc.'},
+                    {value: 'instant_search_price_desc', label: 'Price desc.'}
                   ]
                 "
               >
@@ -65,6 +59,12 @@ import { Component, OnDestroy } from '@angular/core';
   styles: [],
 })
 export class SearchComponent implements OnDestroy {
+  config = {
+    searchClient: algoliasearch('latency', '6be0576ff61c053d5f9a3225e2a90f76'),
+    indexName: 'instant_search',
+    routing: true,
+  };
+
   ngOnDestroy() {
     console.log('SearchComponent::ngOnDestroy');
   }